Add eslint. Refactor installer
